requirements.txt
Django==1.7.7
django-crispy-forms
django-braces
django-debug-toolbar

